\name{create_boxplot}
\alias{create_boxplot}
\title{ Image boxplot }
\description{
  Créer une image d'un boxplot pour chaque colonne d'une matrice dans un fichier.
}
\usage{
create_boxplot(data, output_name)
}
\arguments{
  \item{data}{\code{data} Data frame ou matrice }
  \item{output_name}{\code{output_name} Chemin et nom du fichier image à créée}
}
\author{Edouard Hirchaud }

\keyword{image}
\keyword{boxplot}
\keyword{description}
