\name{MadPro-package}
\alias{MadPro-package}
\alias{MadPro}
\docType{package}
\title{
MadPro : Outils d'analyse de MicroArray pour BIRD
}
\description{
Fonctions permetant la realisation de pipeline d'analyse pour microArray.
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab MadPro\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2011-05-10\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Edouard Hirchaud, Audrey Bihouee, Raluca Teusan, Emeric Dubois,Gerard Ramstein Remi Houlgatte
Maintainer:Edouard Hirchaud  <edouard.hirchaud@univ-nantes.fr>
}
\references{
}
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
